changed the arduino code to support cleaner current transitions
Made the catheter gui a ros-package to leverage rosdocs and roslint
