+++ src/catheter_arduino_gui/.gitignore
@@ -1,1 +1,2 @@
 build/
+doc/
