+++ src/catheter_arduino_gui/include/catheter_arduino_gui/serial_sender.h
@@ -1,18 +1,35 @@
+/*
+  Copyright 2017 Russell Jackson
+
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+*/
+
 #pragma once
-#ifndef CATHETER_SERIAL_SENDER_H
-#define CATHETER_SERIAL_SENDER_H
+#ifndef CATHETER_ARDUINO_GUI_SERIAL_SENDER_H
+#define CATHETER_ARDUINO_GUI_SERIAL_SENDER_H
 
 #include <boost/asio.hpp>
 
 #include <vector>
 #include <string>
 
-#include "ser/simple_serial.h"
-#include "com/catheter_commands.h"
+#include "catheter_arduino_gui/simple_serial.h"
+#include "catheter_arduino_gui/catheter_commands.h"
 
 
 #define PAD_CMDS false
-#define MAX_PAUSE_MS 2000 //ms to wait maximum before declaring timeout, i.e. for sending reset
+// ms to wait maximum before declaring timeout, i.e. for sending reset
+#define MAX_PAUSE_MS 2000
 
 // This file defines the serial port interface. (mid-level)
 void printComStat(const comStatus& statIn);
@@ -78,6 +95,7 @@ public:
 };
 
 
-std::string comStat2String(const comStatus& );
-#endif
+std::string comStat2String(const comStatus&);
 
+
+#endif  // CATHETER_ARDUINO_GUI_SERIAL_SENDER_H
