+++ src/catheter_arduino_gui/package.xml
@@ -48,7 +48,7 @@
   <build_depend>roslint</build_depend>
   <build_depend>boost</build_depend>
   <build_depend>wxwidgets</build_depend>
-  <!--<build_depend>GTest</build_depend> -->
+
   
 
 
@@ -56,23 +56,8 @@
   <run_depend>boost</run_depend>
   <run_depend>wxwidgets</run_depend>
 
-  <!-- <run_depend>GTest</run_depend> -->
-
-  <!-- <run_depend>Boost</run_depend> -->
-  <!--<run_depend>wxWidgets</run_depend> -->
-  <!-- <run_depend>GTest</run_depend> -->
-  
-  
-
-  
-  
-
   <!-- The export tag contains other, unspecified, tags -->
   <export>
-    <!-- You can specify that this package is a metapackage here: -->
-    <!-- <metapackage/> -->
-
-    <!-- Other tools can request additional information be placed here -->
-
+    <rosdoc config="rosdoc.yaml" />
   </export>
 </package>
