+++ src/catheter_arduino_gui/rosdoc.yaml
@@ -1,0 +1,4 @@
+- builder: doxygen
+  name: C++ API
+  output_dir: c++
+  file_patterns: '*.c *.cpp *.h *.cc *.hh *.dox'
