+++ src/catheter_arduino_gui/src/gui/status_frame.cpp
@@ -13,8 +13,10 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 */
-#include "gui/status_frame.h"
+#include "catheter_arduino_gui/status_frame.h"
 #include <vector>
+
+
 StatusGrid:: ~StatusGrid()
 {
 }
