+++ src/catheter_arduino_gui/src/gui/status_text.cpp
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 */
-#include "gui/status_text.h"
+#include "catheter_arduino_gui/status_text.h"
 #include <string>
 
 #ifdef _MSC_VER
